Rename gear forms to displays.
