+++ src/components/gear/gear-armor-display/gear-armor-display.html
