+++ src/components/gear/gear-equipment-display/gear-equipment-display.html
