+++ src/components/gear/gear-list/gear-list.html
@@ -24,15 +24,15 @@
 
 <modal-dialog id="gearArmorModal">
     <span slot="modal-header">Item Details</span>
-    <gear-armor-form id="gearArmorForm" slot="modal-body"></gear-armor-form>
+    <gear-armor-display id="gearArmorDisplay" slot="modal-body"></gear-armor-display>
 </modal-dialog>
 
 <modal-dialog id="gearEquipmentModal">
     <span slot="modal-header">Item Details</span>
-    <gear-equipment-form id="gearEquipmentForm" slot="modal-body"></gear-equipment-form>
+    <gear-equipment-display id="gearEquipmentDisplay" slot="modal-body"></gear-equipment-display>
 </modal-dialog>
 
 <modal-dialog id="gearWeaponModal">
     <span slot="modal-header">Item Details</span>
-    <gear-weapon-form id="gearWeaponForm" slot="modal-body"></gear-weapon-form>
+    <gear-weapon-display id="gearWeaponDisplay" slot="modal-body"></gear-weapon-display>
 </modal-dialog>
