+++ src/components/gear/gear-weapon-display/gear-weapon-display.html
